cdk/requirements.txt
aws-cdk-lib==2.121.1
constructs==10.3.0
boto3==1.34.14
